requirements/production.txt
-r base.txt

# Extra stuff required for production like Prod Database interfacing packages
psycopg2
dj-database-url==0.4.1
gunicorn==19.6.0
kombu==3.0.35
django-timezone-field==1.3
git+https://github.com/pinax/django-mailer
seacucumber
supervisor
django-smtp-ssl
coverage
